+++ packages/pluggableWidgets/activity-indicator-native/CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Dark theme icons for Tile and List view.
 
+### Fixed
+- We fixed incorrectly defined a value in the image props.
+
 ## [2.0.0] - 2021-9-28
 
 ### Added