+++ packages/pluggableWidgets/animation-native/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- We added structure mode preview.
+
 ## [3.1.0] - 2022-1-24
 
 ### Added